Add tests for total supply zero scenario
If total supply is zero during reward period, reward tokens
corresponding to the time while that happens would be lost.

It is a very unlikely scenario, and it has an easy workaround: the
account providing the rewards can make sure that doesn’t happen by
staking a tiny amount. But it’s better to be aware of it.